cocoa: Create view/code helper classes
