+++ FigmaSharp.NativeControls/FigmaSharp.NativeControls.Cocoa/CocoaConverter.cs
@@ -25,7 +25,6 @@
 
 using System;
 using System.Text;
-using System.Linq;
 
 using AppKit;
 
@@ -123,202 +122,6 @@ namespace FigmaSharp.NativeControls.Cocoa
 				return builder.ToString();
 			}
 			return string.Empty;
-		}
-
-
-
-		protected NSControlSize GetNSControlSize(NativeControlVariant controlVariant)
-        {
-			if (controlVariant == NativeControlVariant.Small)
-				return NSControlSize.Small;
-
-			return NSControlSize.Regular;
-        }
-
-
-		protected string GetNSFontSizeName(NativeControlVariant controlVariant)
-		{
-			if (controlVariant == NativeControlVariant.Small)
-				return $"{ typeof(NSFont) }.{ nameof(NSFont.SmallSystemFontSize) }";
-
-			return $"{ typeof(NSFont) }.{ nameof(NSFont.SystemFontSize) }";
-		}
-
-		protected nfloat GetNSFontSize(NativeControlVariant controlVariant)
-		{
-			if (controlVariant == NativeControlVariant.Small)
-				return NSFont.SmallSystemFontSize;
-
-			return NSFont.SystemFontSize;
-		}
-
-
-		protected NSFont GetNSFont(NativeControlVariant controlVariant)
-		{
-			if (controlVariant == NativeControlVariant.Small)
-				return NSFont.SystemFontOfSize(NSFont.SmallSystemFontSize);
-
-			return NSFont.SystemFontOfSize(NSFont.SystemFontSize);
-		}
-
-		protected NSFont GetNSFont(NativeControlVariant controlVariant, FigmaText text)
-		{
-			var fontWeight = GetNSFontWeight(text);
-
-			if (controlVariant == NativeControlVariant.Regular)
-			{
-				// The system default Medium is slightly different, so let Cocoa handle that
-				if (fontWeight == NSFontWeight.Medium)
-					return NSFont.SystemFontOfSize(NSFont.SystemFontSize);
-			}
-
-			if (controlVariant == NativeControlVariant.Small)
-			{
-				// The system default Medium is slightly different, so let Cocoa handle that
-				if (fontWeight == NSFontWeight.Medium)
-					return NSFont.SystemFontOfSize(NSFont.SmallSystemFontSize);
-				else
-					return NSFont.SystemFontOfSize(NSFont.SmallSystemFontSize, fontWeight);
-			}
-
-			return NSFont.SystemFontOfSize(GetNSFontSize(controlVariant), fontWeight);
-		}
-
-		protected string GetNSFontName(NativeControlVariant controlVariant)
-		{
-			if (controlVariant == NativeControlVariant.Small)
-				return $"{ typeof(NSFont) }.{ nameof(NSFont.SystemFontOfSize) }({ GetNSFontSizeName(controlVariant) })";
-
-			return $"{ typeof(NSFont) }.{ nameof(NSFont.SystemFontOfSize) }({ GetNSFontSizeName(controlVariant) })";
-		}
-
-		protected string GetNSFontName(NativeControlVariant controlVariant, FigmaText text, bool withWeight = true)
-		{
-			var fontWeight = GetNSFontWeight(text);
-
-			if (controlVariant == NativeControlVariant.Regular)
-			{
-				// The system default Medium is slightly different, so let Cocoa handle that
-				if (fontWeight == NSFontWeight.Medium || !withWeight)
-					return $"{ typeof(NSFont) }.{ nameof(NSFont.SystemFontOfSize) }({ GetNSFontSizeName(controlVariant) })";
-			}
-
-			if (controlVariant == NativeControlVariant.Small)
-			{
-				// The system default Medium is slightly different, so let Cocoa handle that
-				if (fontWeight == NSFontWeight.Medium || !withWeight)
-				    return $"{ typeof(NSFont) }.{ nameof(NSFont.SystemFontOfSize) }({ GetNSFontSizeName(controlVariant) })";
-				else
-					return $"{ typeof(NSFont) }.{ nameof(NSFont.SystemFontOfSize) }({ GetNSFontSizeName(controlVariant) }, { GetNSFontWeightName(text) })";
-			}
-
-            if (withWeight)
-			    return $"{ typeof(NSFont) }.{ nameof(NSFont.SystemFontOfSize) }({ GetNSFontSizeName(controlVariant) }, { GetNSFontWeightName(text) })";
-            else
-				return $"{ typeof(NSFont) }.{ nameof(NSFont.SystemFontOfSize) }({ GetNSFontSizeName(controlVariant) })";
-
-		}
-
-
-		public nfloat GetNSFontWeight(FigmaText text)
-		{
-			string fontName = text?.style?.fontPostScriptName;
-
-			if (fontName != null)
-			{
-				if (fontName.EndsWith("-Black"))
-					return NSFontWeight.Black;
-
-				if (fontName.EndsWith("-Heavy"))
-					return NSFontWeight.Heavy;
-
-				if (fontName.EndsWith("-Bold"))
-					return NSFontWeight.Bold;
-
-				if (fontName.EndsWith("-Semibold"))
-					return NSFontWeight.Semibold;
-
-				if (fontName.EndsWith("-Regular"))
-					return NSFontWeight.Regular;
-
-				if (fontName.EndsWith("-Light"))
-					return NSFontWeight.Light;
-
-				if (fontName.EndsWith("-Thin"))
-					return NSFontWeight.Thin;
-
-				if (fontName.EndsWith("-Ultralight"))
-					return NSFontWeight.UltraLight;
-			}
-
-			// The default macOS font is of medium weight
-			return NSFontWeight.Medium;
-		}
-
-		protected string GetNSFontWeightName(FigmaText text)
-		{
-            // The default macOS font is of medium weight
-            string weight = nameof(NSFontWeight.Medium);
-			string fontName = text?.style?.fontPostScriptName;
-
-			if (fontName != null)
-			{
-				if (fontName.EndsWith("-Black"))
-					weight = nameof(NSFontWeight.Black);
-
-				if (fontName.EndsWith("-Heavy"))
-					weight = nameof(NSFontWeight.Heavy);
-
-				if (fontName.EndsWith("-Bold"))
-					weight = nameof(NSFontWeight.Bold);
-
-				if (fontName.EndsWith("-Semibold"))
-					weight = nameof(NSFontWeight.Semibold);
-
-				if (fontName.EndsWith("-Regular"))
-					weight = nameof(NSFontWeight.Regular);
-
-				if (fontName.EndsWith("-Light"))
-					weight = nameof(NSFontWeight.Light);
-
-				if (fontName.EndsWith("-Thin"))
-					weight = nameof(NSFontWeight.Thin);
-
-				if (fontName.EndsWith("-Ultralight"))
-					weight = nameof(NSFontWeight.UltraLight);
-			}
-
-			return $"{ typeof(NSFontWeight) }.{ weight }";
-		}
-
-
-		protected NSTextAlignment GetNSTextAlignment(FigmaText text)
-		{
-			FigmaTypeStyle style = text.style;
-
-			if (style.textAlignHorizontal == "RIGHT")
-				return NSTextAlignment.Right;
-
-			if (style.textAlignHorizontal == "CENTER")
-				return NSTextAlignment.Center;
-
-			return NSTextAlignment.Left;
-		}
-
-		public string GetNSTextAlignmentName(FigmaText text)
-		{
-			return $"{nameof(NSTextAlignment)}.{GetNSTextAlignment(text)}";
-		}
-
-
-		public NSColor GetNSColor(string colorStyleName)
-		{
-			return themeColors.FirstOrDefault(c => c.styleName == colorStyleName).themeColor;
-		}
-
-		public string GetNSColorName(string colorStyleName)
-		{
-			return $"{nameof(NSColor)}.{themeColors.FirstOrDefault(c => c.styleName == colorStyleName).themeColorName}";
 		}
